refactor: split tasks.ts into modular structure
Split the 2,300+ line tasks.ts file into smaller, focused modules:
- crud.ts: Basic CRUD operations for tasks
- bulk-operations.ts: Bulk create, update, and delete operations
- assignees.ts: User assignment management
- comments.ts: Comment operations
- reminders.ts: Reminder management
- filters.ts: Client-side filter evaluation logic
- validation.ts: Input validation and utility functions
- constants.ts: Error messages and configuration constants
- types.ts: Task-specific type definitions
- index.ts: Main registration and routing logic

This refactoring improves code organization and maintainability while
preserving 100% functionality. All 997 tests continue to pass with no
breaking changes.

Closes #2
